fixed mode number being set as int. fixed ky!=0 case in coldplasma 1d

+++ petram/phys/em2da/em2da_anisotropic.py
@@ -34,7 +34,7 @@ data =  (('epsilonr', VtableElement('epsilonr', type='complex',
                                      suffix =[('r', 'phi', 'z'), ('r', 'phi', 'z')],
                                      default = np.zeros((3, 3)),
                                      tip = "contuctivity" )),
-         ('t_mode', VtableElement('t_mode', type='int',
+         ('t_mode', VtableElement('t_mode', type='float',
                                      guilabel = 'm',
                                      default = 0.0, 
                                      tip = "mode number" )),)
